+++ hyperwallet/__init__.py
@@ -6,7 +6,7 @@ __author__          = 'The Hyperwallet Developers'
 __email__           = 'devsupport@hyperwallet.com'
 __copyright__       = 'Copyright (c) 2017 Hyperwallet'
 __license__         = 'MIT'
-__version__         = '1.1.2'
+__version__         = '1.1.4'
 __url__             = 'https://github.com/hyperwallet/python-sdk'
 __download_url__    = 'https://pypi.python.org/pypi/hyperwallet-sdk'
 __description__     = 'A Python wrapper around the Hyperwallet API'
@@ -20,6 +20,8 @@ from .models import (
     BankCard,                                                            # noqa
     PrepaidCard,                                                         # noqa
     PaperCheck,                                                          # noqa
+    Transfer,                                                            # noqa
+    PayPalAccount,                                                       # noqa
     Payment,                                                             # noqa
     Balance,                                                             # noqa
     Receipt,                                                             # noqa