+++ CHANGELOG.rst
@@ -1,7 +1,22 @@
 Changelog
 =========
 
-1.1.2 (current)
+1.1.5 (current)
+------------------
+
+- Added Layer 7 encryption for Hyperwallet client
+
+1.1.4 (2018-08-03)
+------------------
+
+- Added PayPal account endpoint
+
+1.1.3 (2018-07-05)
+------------------
+
+- Added transfer endpoint
+
+1.1.2 (2018-03-20)
 ------------------
 
 - Added bank card endpoint